chatbot/requirements.txt
# Travel Chatbot Requirements
# Core Django
Django>=4.2.0,<5.0.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0  # For PostgreSQL support
dj-database-url>=2.0.0  # For DATABASE_URL parsing

# Cache
django-redis>=5.2.0
redis>=4.5.0

# Environment
python-dotenv>=1.0.0

# AI/ML
openai>=1.0.0
requests>=2.31.0

# Development
django-debug-toolbar>=4.0.0  # For development debugging

# Production
gunicorn>=21.0.0  # WSGI server for production
whitenoise>=6.5.0  # Static file serving